+++ api/src/main/java/net/automatalib/ts/concept/DeterministicSuffixOutputTS.java
@@ -13,38 +13,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.automatalib.automaton.concept;
+package net.automatalib.ts.concept;
 
+import net.automatalib.automaton.concept.SuffixOutput;
 import net.automatalib.exception.UndefinedPropertyAccessException;
-import net.automatalib.word.Word;
 
 /**
- * Feature for transition systems that compute a <i>suffix-observable</i> output function, i.e., they compute an output
- * containing a part that can be attributed to a suffix of the input.
- * <p>
- * Note that this is a special case of the {@link Output} feature, as
- * {@code computeOutput(input) = computeSuffixOutput(ε, input)}.
+ * A deterministic suffix output transition system is a
+ * {@link DeterministicOutputTS deterministic output transition system} that can produce
+ * {@link SuffixOutput suffix outputs}.
  *
+ * @param <S>
+ *         state type
  * @param <I>
  *         input symbol type
+ * @param <T>
+ *         transition type
  * @param <D>
  *         output domain type
  */
-@FunctionalInterface
-public interface SuffixOutput<I, D> extends Output<I, D> {
-
-    @Override
-    default D computeOutput(Iterable<? extends I> input) {
-        return computeSuffixOutput(Word.epsilon(), input);
-    }
+public interface DeterministicSuffixOutputTS<S, I, T, D> extends DeterministicOutputTS<S, I, T, D>, SuffixOutput<I, D> {
 
     /**
-     * Computes the output for the given suffix from the state reached by the given prefix.
+     * Computes the output for the given sequence of input symbols beginning from the given state.
      *
-     * @param prefix
-     *         the sequence of input symbols for reaching the state from which the output computation should start
-     * @param suffix
-     *         the sequence of input symbols that should be considered for computing the output
+     * @param state
+     *         the start from which the output computation should start
+     * @param input
+     *         the sequence of input symbols
      *
      * @return the computed output
      *
@@ -52,5 +48,6 @@ public interface SuffixOutput<I, D> extends Output<I, D> {
      *         if the computation encountered undefined transitions that would have been required for computing the
      *         output
      */
-    D computeSuffixOutput(Iterable<? extends I> prefix, Iterable<? extends I> suffix);
+    D computeStateOutput(S state, Iterable<? extends I> input);
+
 }